+++ src/MasterProduct/Models/MasterProductVariant.php
@@ -18,7 +18,9 @@ use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Vanilo\Contracts\Dimension as DimensionContract;
 use Vanilo\MasterProduct\Contracts\MasterProductVariant as MasterProductVariantContract;
+use Vanilo\Support\Dto\Dimension;
 
 /**
  *
@@ -64,6 +66,20 @@ class MasterProductVariant extends Model implements MasterProductVariantContract
     public function masterProduct(): BelongsTo
     {
         return $this->belongsTo(MasterProductProxy::modelClass(), 'master_product_id', 'id');
+    }
+
+    public function hasDimensions(): bool
+    {
+        return null !== $this->width && null !== $this->height && null !== $this->length;
+    }
+
+    public function dimension(): ?DimensionContract
+    {
+        if (!$this->hasDimensions()) {
+            return null;
+        }
+
+        return new Dimension($this->width, $this->height, $this->length);
     }
 
     public function hasOwnName(): bool