+++ src/Shipment/Exceptions/InvalidShippingConfigurationException.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the InvalidShippingConfigurationException class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-03-05
+ *
+ */
+
+namespace Vanilo\Shipment\Exceptions;
+
+use RuntimeException;
+
+class InvalidShippingConfigurationException extends RuntimeException
+{
+}