+++ src/Support/Changelog.md
@@ -6,6 +6,7 @@
 ##### 2023-XX-YY
 
 - Added the `DetailedAmount` DTO class (provides a default implementation for the same name interface)
+- Added the `Dimension` DTO class (provides a default implementation for the same name interface)
 
 ## 3.5.0
 ##### 2023-02-23